+++ RAnalogs/RAnEn/src/ForecastsR.h
@@ -1,12 +1,6 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 /* 
  * File:   ForecastsR.h
- * Author: wuh20
+ * Author: Weiming Hu <cervone@psu.edu>
  *
  * Created on January 21, 2020, 3:36 PM
  */
@@ -14,13 +8,37 @@
 #ifndef FORECASTSR_H
 #define FORECASTSR_H
 
-class ForecastsR {
+#include "Forecasts.h"
+#include "Offset.h"
+
+class ForecastsR : public Forecasts {
 public:
     ForecastsR();
     ForecastsR(const ForecastsR& orig);
+    ForecastsR(SEXP sx_weights, SEXP sx_circulars,
+            SEXP sx_times, SEXP sx_flts, SEXP sx_data);
     virtual ~ForecastsR();
-private:
 
+    std::size_t num_elements() const override;
+
+    const double* getValuesPtr() const override;
+    double* getValuesPtr() override;
+
+    void setDimensions(const Parameters & parameters,
+            const Stations & stations, const Times & times,
+            const Times & flts);
+
+    double getValue(std::size_t parameter_index,
+            std::size_t station_index, std::size_t time_index,
+            std::size_t flt_index) const;
+    void setValue(double val, std::size_t parameter_index,
+            std::size_t station_index, std::size_t time_index,
+            std::size_t flt_index);
+
+protected:
+    Offset offset_;
+    Rcpp::NumericVector data_;
+    
 };
 
 #endif /* FORECASTSR_H */