+++ RAnalogs/RAnEn/src/ObservationsR.h
@@ -1,12 +1,6 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 /* 
  * File:   ObservationsR.h
- * Author: wuh20
+ * Author: Weiming Hu <cervone@psu.edu>
  *
  * Created on January 21, 2020, 3:37 PM
  */
@@ -14,12 +8,32 @@
 #ifndef OBSERVATIONSR_H
 #define OBSERVATIONSR_H
 
-class ObservationsR {
+#include "Observations.h"
+#include "Offset.h"
+
+class ObservationsR : public Observations {
 public:
     ObservationsR();
     ObservationsR(const ObservationsR& orig);
+    ObservationsR(SEXP sx_times, SEXP sx_data);
     virtual ~ObservationsR();
-private:
+
+    std::size_t num_elements() const override;
+
+    const double* getValuesPtr() const override;
+    double * getValuesPtr() override;
+
+    void setDimensions(const Parameters & parameters,
+            const Stations & stations, const Times & times);
+
+    double getValue(std::size_t parameter_index,
+            std::size_t station_index, std::size_t time_index) const;
+    void setValue(double val, std::size_t parameter_index,
+            std::size_t station_index, std::size_t time_index);
+
+protected:
+    Offset offset_;
+    Rcpp::NumericVector data_;
 
 };